+++ internal/palette/ansi.go
@@ -1,3 +1,3 @@
 package palette
 
-var ansiHues []float32 = []float32{-1, 0, 120, 60, 240, 300, 180, -1}
+var ansiHues []float64 = []float64{-1, 0, 120, 60, 240, 300, 180, -1}
